requirements.txt
streamlit
tensorflow-cpu
numpy
pillow
pandas
scikit-learn